Only build the STTY-based input-driver on Unix systems.
This means we can now build a Windows binary :rocket:
